pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Dmiriy Yampolkiy <yampolskiydv@gmail.com>"]
description = ""
name = "memgrep"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
boto3 = "^1.26.66"
cryptg = "^0.4.0"
fire = "*"
fuzzysearch = "*"
loguru = ">=0.6.0"
minio = "*"
# torch = {version = "==1.13.1", index = "pypi"}
# torchvision = {version = "==0.14.1", index = "pypi"}
# clip-interrogator = "==0.4.1"
# easyocr = "*"
more_itertools = "*"
numpy = "*"
psycopg2 = "^2.9.5"
pydantic = "*"
pytesseract = "==0.3.10"
python = "^3.10"
python-environ = ">=0.4.54"
python-telegram-bot = "*"
redis = "*"
requests = "*"
sqlalchemy = "^2.0.2"
telethon = "==1.16.1"

[tool.ruff]
# not enabled ANN101, EM, RUF, D, ERA, T20, A, N
extend-ignore = [
  "D107",
  "D100",
  "D104",
  "D203",
  "D213",
  "C417",
  "B008",
  "B019",
  "ANN101",
  "S101",
  "PD901",
  "PD002",
  "PD015",
  "PD009",
  "PD008",
  "RET504",
  "PGH004",
  "PLR2004",
  "E741",
  "E402",
  "S108",
  "PLR0913"
]
extend-select = [
  "C",
  "B",
  "UP",
  "W",
  "I",
  "YTT",
  "S",
  "FBT",
  "C4",
  "T10",
  "ICN",
  "RET",
  "SIM",
  "TID",
  "PD",
  "PGH",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "PIE",
  "BLE",
  "T",
  "COM",
  "DTZ",
  "Q",
  "ICN",
  "ISC",
  "BLE",
  "C90",
  "E",
  "F"
]
ignore = ["E741", "E402"]
line-length = 120
